+++ SpaceWarpMod/KontrolSystemMod.cs
@@ -13,13 +13,13 @@ using KontrolSystem.KSP.Runtime.KSPAddons;
 
 namespace KontrolSystem.SpaceWarpMod {
 
-    [BepInPlugin("com.github.untoldwind.KontrolSystem2", "KontrolSystem2", "0.5.5.3")]
+    [BepInPlugin("com.github.untoldwind.KontrolSystem2", "KontrolSystem2", "0.5.5.4")]
     [BepInDependency(SpaceWarpPlugin.ModGuid, SpaceWarpPlugin.ModVer)]
     [BepInDependency(KSPAddonsModule.FlightPlanAdapter.ModGuid, BepInDependency.DependencyFlags.SoftDependency)]
     public class KontrolSystemMod : BaseSpaceWarpPlugin {
         public const string ModGuid = "com.github.untoldwind.KontrolSystem2";
         public const string ModName = "KontrolSystem2";
-        public const string ModVersion = "0.5.5.3";
+        public const string ModVersion = "0.5.5.4";
 
         private ModuleManagerWindow? moduleManagerWindow;
         private UIWindows? uiWindows;